mudando retorno da função e o comentario

+++ src/main/java/com/projetospring/repository/AlunoRepository.java
@@ -8,9 +8,9 @@ import com.projetospring.entities.Aluno;
 public interface AlunoRepository extends JpaRepository<Aluno, Long>{
 	
 	/**
-	 * Retorna o número de registros que a tabela aluno
-	 * tem com a matricula igual a especificada.
+	 * Retorna o número de alunos que existem no banco
+	 * com a matricula igual a matricula especificada.
 	 */
 	@Query("select count(matricula) from com.projetospring.entities.Aluno a where a.matricula = ?1;")
-	public int findAlunoByMatricula(String matricula);
+	public Long findAlunoByMatricula(String matricula);
 }
